requirements.txt
fastapi==0.103.2
numpy==1.26.4
opencv-python-headless==4.11.0.86
passlib==1.7.4
pillow==10.3.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
uvicorn==0.23.2
